+++ barapost/src/sorter_modules/parallel_FAST5_utwfunc.py
@@ -16,16 +16,20 @@ from src.printlog import printl, printn, getwt, err_fmt
 index_name = "fast5_to_tsvtaxann_idx"
 
 
-def init_paral_utw():
+def init_paral_utw(write_lock_buff, print_lock_buff):
     """
     Function initializes global locks for funciton 'map_f5reads_2_taxann'.
+    :param write_lock_buff: lock for writing to output file(s);
+    :type write_lock_buff: mp.Lock;
+    :param print_lock_buff: lock for printing to console;
+    :type print_lock_buff: mp.Lock;
     """
 
     global write_lock
-    write_lock = mp.Lock()
+    write_lock = write_lock_buff
 
     global print_lock
-    print_lock = mp.Lock()
+    print_lock = print_lock_buff
 # end def init_paral_sorting